Initial commit: Synchronized Alt Text Plugin for WordPress Gutenberg Editor
WordPress plugin that synchronizes the alt text of images between the Media Library and the Gutenberg Editor.

When an image is added or selected in the editor, the plugin checks if the alt text is empty. If it is empty, the alt text is automatically synchronized from the Media Library.

Features:
- Automatically synchronizes empty alt text from the Media Library to the Gutenberg Editor.